Feat: 로그인된 사용자에 따라 라우터 접근 예외 처리 완료 postman 테스트 필요

+++ routes/categoryRouter.js
@@ -1,11 +1,18 @@
 const {Router} = require('express');
 const asyncHandler = require('../utils/async-handler');
 const categoryService = require('../services/categoryService');
+// 현재 사용자가 로그인했는지 체크하는 미들웨어 적용
+const reqUserCheck = require('../middlewares/reqUserCheck');
 
 const router = Router();
 
 // create (bodyData : name)
-router.post('/', asyncHandler(async (req, res) => {
+router.post('/', reqUserCheck, asyncHandler(async (req, res) => {
+    // 접근한 사용자가 is_admin === true 일 경우 가능함.
+    if(req.user.is_admin === false){
+        throw new Error("접근할 수 없는 요청입니다.");
+    }
+    
     const bodyData = req.body;
     // 지정 할 수 없는 nanoid property 는 bodyData 에서 제거
     Reflect.deleteProperty(bodyData, "nanoid");
@@ -39,8 +46,14 @@ router.get('/:nanoid', asyncHandler(async (req, res) => {
 }));
 
 // update (bodyData : name)
-router.put('/:nanoid', asyncHandler(async (req, res) => {
+router.put('/:nanoid', reqUserCheck, asyncHandler(async (req, res) => {
     const {nanoid} = req.params;
+
+    // 접근한 사용자가 is_admin === true 일 경우 수정이 가능함.
+    if(req.user.is_admin === false){
+        throw new Error("접근할 수 없는 요청입니다.");
+    }
+
     const bodyData = req.body;
     // 수정할 수 없는 nanoid property 는 bodyData 에서 제거
     Reflect.deleteProperty(bodyData, "nanoid");
@@ -53,8 +66,14 @@ router.put('/:nanoid', asyncHandler(async (req, res) => {
 }));
 
 // delete
-router.delete('/:nanoid', asyncHandler(async (req,res) => {
+router.delete('/:nanoid', reqUserCheck, asyncHandler(async (req,res) => {
     const {nanoid} = req.params;
+
+    // 접근한 사용자가 is_admin === true 일 경우 제거가 가능함.
+    if(req.user.is_admin === false){
+        throw new Error("접근할 수 없는 요청입니다.");
+    }
+
     const result = await categoryService.deleteById({nanoid});
     if(result.value === "fail"){
         return res.status(404).json(result);
